[test_nn] split hooks test from test_nn
